requirements.txt
customtkinter
feedparser
requests
pyyaml
schedule
beautifulsoup4
python-dotenv
spacy

